+++ src/plugins/job_submit/lua/job_submit_lua.c
@@ -113,10 +113,10 @@ static pthread_mutex_t lua_lock = PTHREAD_MUTEX_INITIALIZER;
  * overwritten when linking with the slurmctld.
  */
 #if defined (__APPLE__)
-extern int accounting_enforce __attribute__((weak_import));
+extern uint16_t accounting_enforce __attribute__((weak_import));
 extern void *acct_db_conn  __attribute__((weak_import));
 #else
-int accounting_enforce = 0;
+uint16_t accounting_enforce = 0;
 void *acct_db_conn = NULL;
 #endif
